Avoid frustum precision issues as much as possible

+++ src/main/java/me/jellysquid/mods/sodium/client/render/viewport/Viewport.java
@@ -4,9 +4,9 @@ import org.joml.FrustumIntersection;
 
 public class Viewport {
     private final FrustumIntersection[] frustums;
-    private final float x, y, z;
+    private final double x, y, z;
 
-    public Viewport(FrustumIntersection[] frustums, float x, float y, float z) {
+    public Viewport(FrustumIntersection[] frustums, double x, double y, double z) {
         this.frustums = frustums;
 
         this.x = x;
@@ -14,10 +14,10 @@ public class Viewport {
         this.z = z;
     }
 
-    public boolean isBoxVisible(float minX, float minY, float minZ, float maxX, float maxY, float maxZ) {
+    public boolean isBoxVisible(double minX, double minY, double minZ, double maxX, double maxY, double maxZ) {
         for (var frustum : this.frustums) {
-            var result = frustum.testAab(minX - this.x, minY - this.y, minZ - this.z,
-                    maxX - this.x, maxY - this.y, maxZ - this.z);
+            var result = frustum.testAab((float) (minX - this.x), (float) (minY - this.y), (float) (minZ - this.z),
+                    (float) (maxX - this.x), (float) (maxY - this.y), (float) (maxZ - this.z));
 
             // early-exit if not inside one of the frustums
             if (!result) {
